reloading atlas and effect

+++ desktop/src/io/piotrjastrzebski/assetreload/desktop/DesktopLauncher.java
@@ -157,12 +157,12 @@ public class DesktopLauncher {
 
 			Files.walkFileTree(path, new SimpleFileVisitor<Path>(){
 				@Override public FileVisitResult visitFile (Path file, BasicFileAttributes attrs) throws IOException {
-					Files.delete(file);
+					if (Files.exists(file)) Files.delete(file);
 					return FileVisitResult.CONTINUE;
 				}
 
 				@Override public FileVisitResult postVisitDirectory (Path dir, IOException exc) throws IOException {
-					Files.delete(dir);
+					if (Files.exists(dir)) Files.delete(dir);
 					return FileVisitResult.CONTINUE;
 				}
 			});
